Move all custom types to types directory
