+++ src/dataprep/mod.rs
@@ -1,4 +1,4 @@
 pub mod cleaning;
 pub mod ingestion;
 pub mod processing;
-mod subtitle_set_builder;
+pub mod subtitle_set_builder;
